Add Policy to Allow Quota Updates
A user with the 'manager' role should be able to set quotas in dynamic
projects that it manages.
