+++ .gitignore
@@ -1,3 +1,4 @@
 /dist
 /python_unikorn_openstack_policy.egg-info/
 __pycache__
+*.swp
