+++ unikorn_openstack_policy/__init__.py
@@ -12,31 +12,4 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-"""
-Base library entrypoints.
-"""
-
-import itertools
-
-from oslo_config import cfg
-from oslo_policy import policy
-
-from unikorn_openstack_policy import network
-
-def list_rules():
-    """Implements the "oslo.policy.policies" entry point"""
-
-    return itertools.chain(
-        network.list_rules(),
-    )
-
-
-def get_enforcer():
-    """Implements the "oslo.policy.enforcer" entry point"""
-
-    enforcer = policy.Enforcer(conf=cfg.CONF)
-    enforcer.register_defaults(list_rules())
-
-    return enforcer
-
 # vi: ts=4 et:
